+++ datacube/index/postgres/tables/_storage.py
@@ -5,8 +5,7 @@ Tables for indexing the storage of a dataset in a reprojected or new form.
 (ie. What NetCDF files do I have of this dataset?)
 """
 from __future__ import absolute_import
-
-from sqlalchemy import ForeignKey, UniqueConstraint, BigInteger
+from sqlalchemy import ForeignKey, UniqueConstraint, BigInteger, SmallInteger
 from sqlalchemy import Table, Column, Integer, String, DateTime
 from sqlalchemy.dialects import postgres
 from sqlalchemy.sql import func
@@ -18,7 +17,7 @@ from . import _dataset
 # -> Serialises to 'storage_config.yaml' documents
 STORAGE_TYPE = Table(
     'storage_type', _core.METADATA,
-    Column('id', Integer, primary_key=True, autoincrement=True),
+    Column('id', SmallInteger, primary_key=True, autoincrement=True),
 
     # The storage "driver" to use: eg. 'NetCDF CF', 'GeoTIFF'...
     Column('driver', String, nullable=False),
@@ -40,7 +39,7 @@ STORAGE_TYPE = Table(
 # Map a dataset type to how we will store it (storage_type and each measurement/band).
 STORAGE_MAPPING = Table(
     'storage_mapping', _core.METADATA,
-    Column('id', Integer, primary_key=True, autoincrement=True),
+    Column('id', SmallInteger, primary_key=True, autoincrement=True),
 
     # The storage type to use.
     Column('storage_type_ref', ForeignKey(STORAGE_TYPE.c.id), nullable=False),
@@ -84,7 +83,7 @@ STORAGE_MAPPING = Table(
 
 STORAGE_UNIT = Table(
     'storage_unit', _core.METADATA,
-    Column('id', BigInteger, primary_key=True, autoincrement=True),
+    Column('id', Integer, primary_key=True, autoincrement=True),
     Column('storage_mapping_ref', None, ForeignKey(STORAGE_MAPPING.c.id), nullable=False),
 
     # The collection it belongs to.
